requirements.txt
# Analion — Analysis Engine
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
